RLT/RLT/requirements_08.txt
transformers==4.51.1
datasets==3.2.0
accelerate==1.4.0
bitsandbytes==0.45.5
peft==0.14.0
hf-transfer==0.1.9
deepspeed==0.17.2
trl==0.14.0
hydra-core==1.3.2
ray==2.47.1
matplotlib
wandb
tensorboard==2.18.0
fire==0.7.0
